move contributors, redirect index to /teams
