+++ next.config.js
@@ -16,6 +16,15 @@ const nextConfig = {
         ],
       },      
     ]
+  },
+  async redirects() {
+    return [
+      {
+        source: "/",
+        destination: "/teams",
+        permanent: true
+      }
+    ]
   }
 }
